pyproject.toml
[project]
name = "bookingagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "anthropic>=0.75.0",
    "dotenv>=0.9.9",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-api-python-client>=2.108.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "langgraph>=0.0.40",
    "langchain>=0.1.20",
    "langchain-groq>=0.1.5",
    "groq>=0.37.1",
]
